+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Handle user revoking video input permission
+- Fix FinishGatheringICECandidatesTask when there are no turn credentials
 
 ## [1.7.0] - 2020-05-23